Changed Script.instructionCount to uint16 (was uint8)

+++ c++/UziFirmware/Program.cpp
@@ -97,7 +97,7 @@ Error readScripts(Reader * rs, Program* program)
 	program->scripts = uzi_createArray(Script, program->scriptCount);
 	if (program->scripts == 0) return OUT_OF_MEMORY;
 
-	uint8 instructionCount = 0;
+	uint16 instructionCount = 0;
 	for (int16 i = 0; i < program->scriptCount; i++)
 	{
 		Script* script = &program->scripts[i];
